+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ----------
 
 v0.14.22 (??.06.2015)
+ * new getLastModifyDateTime and getCreationDateTime methods for base model
  * new getSysLanguageUid and getLabel methods for base model
  * optional options and overlay parameters for tx_rnbase_mod_Util::getSpriteIcon